refactor pending actions
